# docs/swagger.yaml
definitions:
  entity.LoginResponse:
    properties:
      message:
        type: string
      token:
        type: string
    type: object
  entity.Task:
    properties:
      createdAt:
        type: string
      description:
        type: string
      done:
        type: boolean
      finishedAt:
        type: string
      id:
        type: string
      ownerId:
        type: string
      title:
        type: string
    type: object
  entity.User:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      password_hash:
        type: string
    type: object
  handler.Message:
    properties:
      message:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/health-check:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Message'
      tags:
      - General
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Return auth token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Message'
      tags:
      - Auth
  /api/v1/auth/signup:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Message'
      tags:
      - Auth
  /api/v1/task/:
    get:
      consumes:
      - application/json
      description: Return all tasks of current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Task'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Message'
      tags:
      - Task
  /api/v1/task/{taskId}:
    delete:
      consumes:
      - application/json
      description: Exclude task
      parameters:
      - description: The ID of Task
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Message'
        "500":
          description: Internal Server Error
          schema:
            type: string
      tags:
      - Task
    get:
      consumes:
      - application/json
      parameters:
      - description: The ID of Task
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Message'
      tags:
      - Task
    post:
      consumes:
      - application/json
      description: Update task
      parameters:
      - description: The ID of Task
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Message'
      tags:
      - Task
  /api/v1/task/{taskId}/finish:
    post:
      consumes:
      - application/json
      description: Set status of task with finished
      parameters:
      - description: The ID of Task
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Message'
        "500":
          description: Internal Server Error
          schema:
            type: string
      tags:
      - Task
  /api/v1/task/create:
    post:
      consumes:
      - application/json
      description: Register new task and return data
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Message'
      tags:
      - Task
  /api/v1/user/me:
    get:
      consumes:
      - application/json
      description: Return current user info
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Message'
      tags:
      - User
swagger: "2.0"